Rename options elements
